refactor: rename cf module to conda_forge for readability
Group standards prefer maximal human readability in names. Python
modules cannot contain hyphens, so conda_forge is used rather than
conda-forge.

Renames src/scikit_package/cli/update/cf.py and its test file, and
updates the import and both call sites in scikit_package_app.py. The
package update conda-forge CLI is unchanged.
