+++ news/simon-cf-robust.rst
@@ -1,0 +1,23 @@
+**Added:**
+
+* <news item>
+
+**Changed:**
+
+* Renamed the internal ``cf`` module to ``conda_forge`` for readability. No change to the ``package update conda-forge`` CLI.
+
+**Deprecated:**
+
+* <news item>
+
+**Removed:**
+
+* <news item>
+
+**Fixed:**
+
+* <news item>
+
+**Security:**
+
+* <news item>
